Sort the tests in each category so they run in alphabetical order instead of arbitrary order
Previously, tests were running in arbitrary order because they were inserted into a List as they were found in the properties, which is a HashTable and thus arbitrarily ordered.
This version uses a SortedSet instead of a List, where the sort order is defined using the test name (ConfigurationTest now implements Comparable).
It might be better if the tests ran in the order they are specified in the category.