Changed value for param "activation-mode" from "base" to "base-level" to match csoar
Since we're doing params with enums and enums can't have dashes, had to customize the enum a bit and change how we're calling it.
Also caught the illegal arg exception and rethrew it as a soar exception so a proper error message gets printed to the trace (just like csoar).