Switched epmem and smem commands to use custom "on/off" enums instead of boolean providers for parameters
This fixed a few bugs where the commands took "true/false" instead of "on/off" and also outputs an error message if an invalid setting is passed into set (as opposed to defaulting to false).
Also catch illegal arg exception in epmem and rethrow as soar exception so the error message shows up in the debugger properly.