Undid part of commit 07ff79c0510a823e002bf8c34e99923b671af67e in order to make source work with spaces in path
That commit claimed that it was fixing a problem with spaces in source commands, but it appears to have caused this more significant problem.
My specific case was sourcing a file that sourced another file using a relative path that included a lot of "../" such that, when expanded to an absolute path, it contained spaces.