Don't have debugger show/track justifications in production table. In long runs with lots of subgoals, they end up accounting for huge amounts of memory. Also, most people don't think of them as rules, so seeing them in this list would be confusing.
It appears that for some reason, while all justifications get added to the table, not all of them get removed. As near as I can tell, this is because the event for production removal isn't triggered for all justification removals (but I could be wrong).
Essentially, after a long run, it appears that there are 1000's of justifications being stored in this model, but there shouldn't be nearly that many in the system (justifications tend to be short-lived).