Fixed issue where couldn't load URLs from jars that had "." in their path. Added tests for default and tcl interpreters.
This came up in an older version of NGS. The fix is to normalize the paths before trying to source them. The standard approach is to convert to a URI and call the normalize method. Of course, URI normalization doesn't support jar paths. So I added UrlTools.normalize to handle jar paths (and do standard normalization for everything else).