Fixed unused variable/method warnings.
For variables that were set and updated, I just suppressed the warning (the fact that they are not currently used may indicate missing functionality from csoar). For methods/variables that were really not referenced anywhere, I removed.
For the SyntaxPatternComponent, I changed code so the member variable was used instead of the constructor parameter. Arguably this doesn't need to be a member var at all, but doing it this way may make it easier to refactor the code in the future (e.g., move this stuff out of the constructor).