Removed a line from a test that doesn't actually do anything (but was causing a warning).
Note that if I fix this line so it does the removal, the test fails. Additionally, I'm pretty sure this line is not intended, as it doesn't seem to make sense to remove an element from a list while you are looping over that list -- you would eventually run off the end of the list or get a concurrent modification exception (even if the test didn't assert for other reasons).