Fixed RunCommand and RunCommandTest. 0 warnings in entire system!
Changed RunCommand to be more similar to how it was before the picocli update: only takes an AgentRunController in the constructor. Because it doesn't have an agent, it instead captures the output to a string (via new Utils methods I added for this), and returns the result via the normal SoarCommand mechanism (instead of returning an empty string). As part of this, the tests had to be updated to check the printed result instead of looking for thrown exceptions -- I just check that the help got printed, which indicates an error. Finally, the tests revealed a couple minor things that didn't get translated from the original correctly (e.g., it now errors if multiple run types are specified, or a count of 0 is specified).