Fixed smem tests (smem command doesn't directly return results anymore, but instead prints directly to the trace). Switched smem command to throwing exceptions when there are errors. Don't start output with newlines (those are added elsewhere in the trace, and the newlines causes some tests to fail).
Note that the difference between directly returning results and printing them directly to the trace is that directly returned results can be stored in a tcl variable directly (e.g., set x [myCommand]), but results printed directly to the trace cannot. It appears that previously lots of jsoar commands returned results, but csoar commands rarely do (pwd is an exception), so changing this in jsoar increases csoar compatibility (i.e., someone won't write tcl code in jsoar and then be surprised that it doesn't work in csoar).