Add WIP bundled debugger
Still has some bugs/rough spots to fix, but should be reasonable usable.
Pushing upstream to make sure it doesn't get lost in the weeds.

Note that this does bump up the target Android SDK a few releases (26
and minimum 22 aka 5.1.1 from 2014), but hopefully not beyond what is
realistic targeted anyway.  Doubt a device old enough to not support
5.1.1 has the horsepower to run JSoar anyway.