Removed pwatch and production-watch commands. Moved production-watch to "production watch". Replaced calls to pwatch in tests with "production watch" (the new command). Removed extra newline in one bit of output that was making a test fail. Production command now prints directly to trace rather than returning a result, so capture output there for test.
I will eventually add appropriate aliases for the old commands, but I don't want to add testing aliases to a test that wasn't previously using aliases, which is why I changed the command instead.
The pwatch command was reimplemented into production-watch instead of "production watch" due to a typo in the confluence document describing the mapping between old and new commands, so that's fixed now, too.