Ensured all interp.eval's happen in one place and are thread safe.
This fixes an issue where sometimes, due to threading that I don't fully understand, we would call interp.eval a second time before we retrieved the result from the first interp.eval. This would invalidate the first result and throw an exception. Now everywhere calls SoarTclInterface.eval, which is the only place that does interp.eval, and that method is synchronized, so the result doesn't get invalidated before it's been retrieved. I can't guarantee this fixes the issue, but before it was easily reproducible, and now it's not.