Undo change from commit 311ea9e8e2f576890cbcea86551b548eedcaf16f that broke PrinterTest.testSpaces().
In the test, 500 spaces are printed 80 at a time. Because the min statement was outside the while loop (which changes the value of n), this function could only print spaces in groups of 80 -- 500 is not evenly divisible by 80, and thus we ended up with an extra 60 spaces. Moving the min back inside the loop allows us to print the last 20 spaces properly.