Add support for "resource:" prefix in the source command, which makes "package require" work.
jtcl's native source command supports this, but we replace it with jsoar's source command, which doesn't. But it turns out that treating it like a "classpath:" prefix works, so we just internally swap the prefixes. This actually isn't specific to the tcl interpreter (you could use resource: in the default interpreter, too).