Show the horizontal scrollbar in the trace again.
There appears to be some kind of built-in line wrapping, but it occurs at a width that is much larger than the displayed width. So the horizontal scrollbar is to help fix that. Making it always show to prevent a potentially expensive calculation to determine if it should show or not.
Note that we used to use proper line wrapping when we were using JTextArea, which has it built in, but it seems to get similar behavior with JTextPane will require custom overrides.