Changed loading of default tcl code to not utilize the built-in evalResource, which doesn't appear to work in Java 11 projects that use modules
The issue seems to be that the default tcl code resource is in a different module than jtcl, and thus the resource is not accessible from jtcl. Since jsoar is not a Java 11 project itself, it can't explicitly open the resources either. The workaround is to manually get the URL from inside SoarTclInterface, and then pass that to jtcl's interp.evalURL