Treat all unrecognized options as parameters
Previously this was only done for the debug command, because we had run into specific issues there. But it turns out it can come up in many other places, too (e.g., srand and log can both take negative numbers as parameters that otherwise get treated as unrecognized options). So we're applying this behavior to everything, and we'll see if it causes any problems between now and the next non-snapshot release.