Fixed issue where the debugger trace view was using a persistent writer instead of a normal writer
It appears this changed during debugging at some point, so I changed it back. The issue was that when something tried to prevent writing to the trace by pushing a nullwriter, it would still show up there (and sometimes twice, as things were written to the normal writer and the persistent writer). E.g., ngs4's CORE_GetCommandOutput.