Updated PicocliSoarCommand so the agent is optional. Updated pwd back to not needing an agent.
There are some commands for which an agent is unnatural. E.g., pwd really wants a SoarCommand, which ultimately is associated with an agent, but the agent is not accessible. If we had to separately supply an agent, we could end up in a situation where two different agents were accidentally used, which could only be bad. By making PicocliSoarCommand more flexible, we avoid this issue.