Removed unnecessary graal dependencies
These other dependencies could be needed if we were to switch to the Graal API for scripting or we wanted to enable runtime JIT. Those are things we may do in the future, when the other languages are better supported in graal.
Note that technically the chromeinspector dependency is not necessary, but including it enables additional logging which can be helpful when debugging (e.g., it will print when the js engine is initialized).