Updated to latest picocli and fixed log command to work with breaking change
The breaking change is that strings that look like options can no longer be used as arguments to options without quoting them. The log command supported "log on" and also "log --echo on" -- the second one stopped working because of the first one. Given that there are a lot of other ways to turn on logging (including log --on), on and off are the only option that don't require the "--" prefix, and this command doesn't exist in csoar so there are no compatibility concerns there, I just took the naked on/off options out. So you have to use the "--" versions now.