Added java 11 module names to manifest -- this maintains java 8 compatibility
An issue with using java 8 libraries in java 11 applications is that the java 8 library ends up with an automatic module name based on the file/directory names, and that causes warnings to be generated on the java 11 side, since the module name is considered "unstable". This solution puts the module name info into a manifest file, which satifies the java 11 side while also avoiding the creation of a module-info.java file, which isn't java 8 compatible.

Technically this breaks compatibility for Java 11 applications since the new module name, org.jsoar.core is different than the previous automatic module name, jsoar.core. So we'll have to increase the major version number, even though there have been no functional changes.