Updated to use jackson to read yaml into beans, rather than using a custom deserializer
Years ago, the original thought was to make the file "user friendly" by not exposing the direct structure of the underlying object there, and instead using "friendly" names for the fields. Since then, I've changed my mind. I think that the typical user of this will be very technical and not at all put off by camelCase field names, and the reduced code means it will be much more maintainable, as there's not much interesting going on (the exception being the way the default values are handled, but even that is not complicated; see https://stackoverflow.com/questions/65483376/deserialize-yaml-in-java-using-defaults-read-earlier-in-file/).