Fixed csoar testing
Some of the example tests seem to get stuck in csoar, presumably because of changes to smem/epmem/chunking since the tests were originally written. But the tests that don't use that stuff seem to work fine.
One of the major changes was the removal of the hack to set java.library.path at runtime. It seems to just work to set the system property now (but the hack does not work). It may be that this is a jvm-specific issue -- I'm using Amazon Coretto JDK 15, but perhaps an Oracle JDK would require the hack.