Probably fixed race condition where autocompletion was happening on the swing thread instead of the Soar agent thread
The issue here is that autocompletion involves interacting with the Soar commands. The Soar commands reuse the same PicocliSoarCommand instance for performance reasons. Because the debugger is simultaneously executing the command and attempting to autocomplete, the same PicocliSoarCommand instance was being interacted with by two different threads at the same time -- badness insued.