Increased test timeout a lot
The reason for increasing the timeout so dramatically is that the timing of the sleeps on github is very unreliable (I've seen a 500 ms sleep take almost a full minute). This is probably related to the number of threads being created and the load the system may be under. Hopefully this timeout will be sufficient, while still casuing the test to fail within a somewhat reasonable time if it really is stuck.