Ensure that the AgentThread stops when explicitly interrupted
This happens when a ThreadedAgent is detached. But if the agent happens to be in the middle of eval'ing something and the Tcl interpreter is being used, then jtcl may swallow the InterruptedException and prevent the AgentThread from stopping. To address this, I added a separate interrupted flag that we set, so even if the InterruptedException gets swallowed, we will still stop the AgentThread. This will not help in cases where an InterruptedException is thrown for some reason other than an explicit interrupt and gets swallowed, but I don't know of any cases where that happens.