Ensure that any exceptions printed to the trace are prefixed with "Error"
This might be redundant with some existing exception messages, but by ensuring we always prefix with Error, we can do things like check the trace for errors. Soar always swallows these exceptions at this point, so if we want to know if there was an error, checking the trace seems to be the way to do it.