Removed tcl interp dispose call as it can be extremely expensive
Disabling this code on a trial basis because it can make agent deletion extremely expensive if there are a lot of Tcl objects (e.g., if using incr tcl). As far as I can tell, since this class is going away immediately and there's no way for anything outside this class to reference the interp, this should be safe. Basically, internally Tcl is iterating through all the objects (recursively) to make sure there are no other references to them so they can actually be cleaned up. For what we're doing, I don't think this can actually happen.