Updated SoarUnit command line output to make the summary easy to read
Previously when outputting the results SoarUnit would include the traces from any failed tests in the middle of the summary. This meant it was really hard to actually see the summary results -- you could see the total passed/failed at the very end, but seeing which tests passed/failed was buried in all the trace output. The output now shows the traces of the failed tests first, and then puts the per-test summary without traces at the end. This is the same output as before, but no longer buried in the trace output so it's actually useful.