When executing a command, hide completions and help tooltip
Basically, when the user presses enter, we can dismiss these elements as they cover parts of the UI, and they aren't needed since the user has finished their command. To achieve this, we have to track when we're executing a command so that the various events that fire when the user types don't cause the completions and help to show up like they normally would. We also removed the delay in showing the completions and help -- this was causing complications where these would be shown after a command was executed because they were invoked during typing, but with a delay. The delay is of questionable value anyway, so it was removed.