Fixed issue where pressing enter in completions list did not put the selected completion in the command box, but instead tried to execute the partial command in the command box
For reasons I don't understand, the previous code, which bound the enter key to a function that would do this, never executed. It seems that the editor's action listener fired instead. So I removed the action listener and put a keybinding on the editor that handled both cases. (Two separate keybindings, one on the edtor and one on the completionsList also did not work.)