Fixed bugs in qmemory command where value wasn't reset between calls
It's possible this is a bug in picocli, but the structure of the command was kind of strange -- the value only needed to be associated with the set option, but it was a separate which led to some strange behavior. As part of this, I also removed unnecessary fallback behavior, like assuming that a value with no command meant you wanted to get the value. That sort of thing is just an error now, and it's handled by picocli vs. qmemory doing its own checking.