Added logging implementation to application projects
Since these are actually meant to be executed as applications, it makes sense to provide an actual logger. I decided not to do this for jsoar-repl, even though it is an application too, since it has a command line interface and it would probably be really distracting to see log output in the middle of your repl.