Fixed warnings about various types not being visible to modules that would require these modules
There were two kinds of issues. The issue is basically when a public exported method refers to some type (as a return value or parameter) that is not public and/or exported, or comes from a module that is not required. To fix the second (and most common) issue, I made a bunch of requires transitive, so they would bring in all the other requires needed to use that module. For the first issue, I made a bunch of stuff public that was previously package private. In both cases I could have instead limited what got exported so that the functions in question weren't even exposed, and thus it wouldn't have been an issue. But I don't have a good reason yet to limit access to the internals -- I'd rather expose everything so users can do whatever they want, than hide something that later turns out to be needed.

Ok, there was one last issue where a test was giving the not exported warning, but I suppressed that because test classes are not typically exported.