Removed unnecessary resources entries in pom.xmls
As of maven 3.9.0, these seem to actually cause problems. Specifically, there was a malformed input exception when trying to filter resources for jsoar-performance-testing -- but nothing needs to be filtered. I suspect these were left over from the original port to maven, when things did not necessarily follow the maven defaults as far as source layout goes.