Removed code that saved some properites in the rete
This was never correct; it deviates from the csoar implementation and could result in unexpected behavior. The correct way to do this is to use a .soar file to load the rete and set whatever settings you want to set. Ironically, the tests relied on this behavior, so I had to update the tests to restore certain properties manually. This was only necessary because the tests are not setup to make using a rete easy (they don't use a separate file for settings that can be loaded with a rete).