Updated eclipse formatter settings and reformatted code
The primary motivation was to fix all the tabs that should have been spaces. As part of this process, I also expanded the max line width from 80 to 200 (monitors are wider these days) and changed some of the formatting settings I didn't agree with (and the code didn't follow anyway) (e.g., I removed the space before paren for a lot of things). Inevitably there is a lot of unnecessary whitespace changes in here, but I didn't see a way to tell the formatter I don't care about certain things and just to leave it as-is (e.g, indentation of blank lines).