Moved uber jar creation to a separate profile
Perhaps this is new as of maven 3.9, but it looks like when the normal and uber jars are deployed, only one pom is deployed and it looks like the one from the uber jar is "winning" now whereas it never did before. This means that attempts to use the normal jar fail because the dependencies are missing from the pom.