Used diamond operator in most places
From sonarqube. I left a few unchanged where it was obvious from the immediate context what the type was, but the vast majority of cases the type is already on the same line, and of the remainder the vast majority are within a few lines (e.g., the return type of a method).