Fixed a bug in autocompletion related to aliased commands
Internally when picocli is generating autocompletions it gets the parent commands. But if we have aliased a subcommand then picocli gets into trouble because the parent command exists but isn't fully initialized, leading to an NPE. The fix here is to check on our side if there are parent commands and if there are, to construct the full (unaliased) command to send to picocli.
Additionally, there was some unnecessarily complex logic to determine the argIndex and positionInArg parameters that picocli needs, so that has been greatly simplified.
Finally, we were passing the cursorPosition around but not using it for anything, so I removed that argument.