Fixed warning about a type not being available to downstream users
The method RestletTools.setResponseHeader takes an argument of type Response, which comes from org.restllet. By transitively requiring the module, we ensure that type is available to users of that method. Practically this is an application and not really a library, so it's unlikely anyone else would need access to this, but it doesn't hurt to do this.