Added service files back in to more easily support uber jars and non-modular applications
While these aren't needed by jsoar itself, as the information is already specified in the module-info.jar files, that information is lost when building a uber (i.e., shaded) jars, and is not available to non-modular applications. Previously I said that non-modular applications should just add these files to their own resources manually, but it turns out we can't even easily build a standalone jsoar debugger (or soarunit, or repl) without adding these files to those projects, too. At that point, I decided just to put them back in jsoar-core so they are available to everyone who needs them. They are still specified in the module-info.java as well. This may become a maintenance issue as now things need to be updated in two places, but those files haven't been touched in many years, so I'm not too concerned about it.