Updated dependencies
Of particular interest here is the guava update -- when used in a gradle build (e.g., for android) the new version will automatically select the correct version (jre vs. android) depending on what the target is, regardless of which version is actually specified in the dependencies. That said, it appeared to be working just fine on android with the jre, but this may help avoid possible issues in the future.