If no execution time source is found via the service loader, use a default
This is to work around an issue in some old versions of Android plugins where the META-INF in libraries gets stripped out, which means the services in jsoar won't load. All other services have fallbacks such that nothing crashes except this one, so I put a fallback in. To my knowledge this is the only execution timer anyone has ever used anyway.