Use guava android version for maximum compatibility. Fix the couple places that used something not available in the android version.
In the past we would just use the jre version on android, and since we only care about more recent android versions that support Java 8 features, this was fine. But guava now automatically detects when you're targeting android with gradle and switches to the android version no matter what you selected. This means if the code is using something not in the android version (e.g., the Streams class) then you get a runtime crash because it can't find the class. There doesn't seem to be a great way to disable this switch right now, but restricting ourselves to the android version is not a big deal (for now). We don't directly build jsoar in this way, but other apps that have jsoar as a dependency do work this way and hence jsoar is causing problems in those cases, even if they aren't using the affected RHS functions.