Major reworking of token parsing code to support tokens with spaces
Tokens with spaces are delimited by pipes. This code is not perfect because without starting from the beginning of the entire text we can't be sure whether a click is inside or outside a pipe. We make the simplifying assumption that it's ok to just start at the beginning of the line (which is probably correct 99.9% of the time). Then we walk forward, tokenizing until we get to the token that contains the clicked character. This appears to work well, and there are even tests now :)