Updated dependencies
Note that this does not work on Java 24 (and possibly 22 or 23, but didn't test those). The issue is that graalvm js uses an unsafe method that has been removed. Updating graalvm js will require changing the minimum supported Java to 17, as this is required by the graalvm versions that fix the issue.